+++ DirectX11/graphics/IndexBuffer.cpp
@@ -5,7 +5,7 @@
 #include "IRenderingContext.h"
 
 IndexBuffer::IndexBuffer(IRenderingContext& graphics, const std::vector<USHORT>& indices)
-    : indexCount(indices.size())
+    : indexCount(static_cast<UINT>(indices.size()))
 {
     D3D11_BUFFER_DESC desc
         {
