+++ DirectX11/graphics/TransformationBuffer.cpp
@@ -6,7 +6,7 @@
  std::shared_ptr<TransformationBuffer> TransformationBuffer::MakeFromTransform(IRenderingContext& graphics, const GioTransform& transform)
 {
     DirectX::XMMATRIX matrix = TransformUtils::CalculateFinalMatrix(graphics, transform);
-    return std::make_shared<TransformationBuffer>(graphics, &matrix, sizeof(matrix));
+    return std::make_shared<TransformationBuffer>(graphics, &matrix,static_cast<UINT>(sizeof(matrix)));
 }
 
 TransformationBuffer::TransformationBuffer(IRenderingContext& graphics, const void* data, UINT dataWidth)
